PP-12358 Combine two test classes
